Build against several libheif versions
